+++ CHANGELOG.md
@@ -229,3 +229,4 @@ to `cache_dir`. `saving_dir` is now deprecated and will be remove in version 0.8
 - Improve doc
 - Add a note for parsing data cleaning (i.e. lowercase, commas removal, and hyphen replacing).
 - Add hyphen parsing cleaning step to improve Canada address parsing (see [issue 137](https://github.com/GRAAL-Research/deepparse/issues/137)).
+- Add ListDatasetContainer for Python list dataset.